requirements.txt
flask
opencv-python
pytesseract
pandas
openpyxl
requests
easyocr
Pillow 